Update Download_GEFSv12_function.py
I have modified the function `Download_GEFSv12_function` so that it can download variables at hybrid/pressure levels.
The argument `lev` is no longer restricted to be a scalar, for data at hybrid/pressure levels it can be a list of the levels that are desired (e.g., `[1,2,3,4]` for hybrid levels, or `[1000,900,850]` for pressure levels). Also, `lev` now contains integers (or `None` if there is no specified level), while is was previously containing strings.