Update Download_GEFSv12_function.py
I have simplified the function a bit so that it doesn't loop through the variables, as this loop is in the batch script already. 
Previsouly, it was actually looping through `['Days:1-10','Days:10-16']`, but now this is passed as an argument of the function (`resol`), so that there is no loop over `v` in the function.
The delineations of the area, `bounds` are now also passed as argument instead of being hard-coded in the function.