Update Download_GEFSv12_function.py
Added argument `pathExstFiles` to the function. 
If provided (default is `None`), the function tries to find the grib file in that location. If it can't be found, then the file is downloading from AWS.
I followed Lesley's indications about how the files are stored on the PSL archive:
On the AWS they have 2 files per date with the same name but different subdirectories. On PSL we changed this so every file has a unique name.
For example, on AWS, there are files:
2000010500/c00/Days:1-10/ugrd_hgt_2000010500_c00.grib2 and
2000010500/c00/Days:10-35/ugrd_hgt_2000010500_c00.grib2
but on PSL the same files are 2000010500/c00/ugrd_hgt_2000010500_c00_day1-10.grib2 and
2000010500/c00/ugrd_hgt_2000010500_c00_day10-35.grib2